# [wip] Enable ThinLTO with incremental compilation.