This PR allows `rustc` to use ThinLTO and incremental compilation at the same time. In theory this should allow for getting compile-time improvements for small changes while keeping the runtime performance of the generated code roughly the same as when compiling non-incrementally.

### How does it work?

- ThinLTO computes the so-called `ModuleSummaryIndex`. This index contains a map of which module imported stuff from which other modules during ThinLTO. 
- The implementation in this PR takes this "import map" after it has been generated and saves it in the incr. comp. cache directory.
- In a subsequent compilation session, we load the previous import map and use it do enhanced cache invalidation: Because ThinLTO pulls in things from other modules we don't only have to re-compile an object file if its corresponding source code has changed but also if something has changed in a module that we imported something from. 
- The cache invalidation logic thus changes from `has_changed(module.inputs)` to `has_changed(module.inputs) || any(has_changed(imported_module.inputs))`.

### Performance

The effect on compile times can be seen here: https://perf.rust-lang.org/compare.html?start=68c39b9fec17846005da9a8e42991422c08c377c&end=02e01bd388a35a90002e37851aa26ba59c593a42&stat=wall-time

Some observations:

- The `baseline incremental-opt` case is always slower. This is expected since we are running ThinLTO now but didn't do so before.
- Some of the `patched incremental: println-opt` cases are *a lot* slower. E.g. `clap-rs-opt` which goes from 4.47s to 24.65s, i.e. it takes more than 5 times as long. The small change in these cases probably invalidates a module that many other modules import things from. 
- Some of the `patched incremental: println-opt` are hardly affected. E.g. `encoding-opt` goes from 0.38s to 0.42s, i.e. 8% slower. It is still 10 times faster than compiling from scratch, in theory without sacrificing runtime performance.
- Some of the `debug` builds are also negatively affected (e.g. `regex-debug / patched incremental: compile one`). I suspect this is because in it's current form, the PR allows for less parallelism during codegen. It first sequentially checks all modules for modifications and only then starts codegen and LLVM. Without this PR, modification checking is done more lazily, while LLVM is already running. This should not be hard to fix for the non-ThinLTO case.

### How to proceed?

- Bring back the performance for the incremental non-ThinLTO builds. This should be done before merging this PR.
- Find out how the heck to test this properly. This would also be good to do before merging.
- Maybe merge the PR and then ask people on irlo to compile their code with it. This way we would get us numbers on the *runtime* performance of code generated this way. It would be good to know if this is even worth the trouble. 
- Once we know that runtime performance is good, we can decide whether to turn incremental compilation on by default also for release builds. Just merging this PR won't do so.
- Investigate if we can make cache invalidation smarter. The ThinLTO import map tells us exactly which functions it imported into a given module, so we could try to check things at the function- instead of at the module-level. I'm sure though that that's not an easy task.

cc @rust-lang/compiler @rust-lang/wg-compiler-performance

~~This (work-in-progress) PR allows for combining ThinLTO and incremental compilation. I'll write up something more detailed once the kinks are worked out. For now, I'm mostly interested in what this does to compile times.~~