@cuviper 2018-07-12
The error is on llvm-emscripten, based on LLVM 4.0 which doesn't have `keys()`.  LLVM 5 does have it.