@michaelwoerister 2018-07-13
I think the results are kind of expected because of how coarse-grained this still is. Right now if something changes in a module then all modules important *anything* from that module (even something that has not really changed) need to completely re-compiled.

There's room for improvement: If we track changes for every `CodegenItem` separately (instead of for whole `CodegenUnits` at a time) then we (probably) can use the more detailed information available in the ThinLTO import map.