@michaelwoerister 2018-07-13
I updated the PR description: https://github.com/rust-lang/rust/pull/52309#issue-201036336

@rust-lang/compiler @rust-lang/wg-compiler-performance