@alexcrichton 2018-07-13
Thanks @michaelwoerister! I wonder though, I think that incremental ThinLTO upstream at least isn't supposed to have such bad performance, so I think we're missing a step in impelmenting this? I thought incremental ThinLTO would looks something like this:

* As usual, incrementally translate and optimize a bunch of LLVM modules. This may have a bunch of cache hits (especially in the println case)
* Next we calculate the ThinLTO module summary index for everything (unconditional work)
* Next we take this index and generate a unique hash for each module which includes items imported from other modules (unconditional work)
* If there is an object file matching this unique hash we use it from the cache, otherwise we generate a new object file via codegen.
* Finally, all the output object files here are linked together

It looks like the implementation here is invalidating too eagerly? The construction above would avoid retranslating/optimizing modules that depend on changed modules but don't import any of the changes. In that the cache invalidation needs to only happen for the ThinLTO passes, not the entire module's original optimization passes.