@alexcrichton 2018-07-13
@michaelwoerister indeed! Although I think what I mentioned above is how incremental ThinLTO was designed to work? In that if you're using a linker plugin then all your intermediate files are the IR files and then the final link also has a cache directory with a lot of object files? I think we have some other various methods to greatly improve our cache directory sizes, for example we should probably be using thin archives which just point to the object files elsewhere on disk, and that way object files aren't copied around. (additionally we probably save way too much IR in too many places!)

Within our own incr. comp. framework I think ThinLTO would roughly look like:

1. There's a query to get the final set of object files. Each object file is the same as our already-existing scheme to partition into CGUs.
2. Each CGU's object file query queries for the pre-ThinLTO-IR of the object file (optimized) as well as the module summary for this object file.
3. Querying the module summary queries for the global module summary and then selectively queries for specific objects from other IR modules. Note that here you don't depend on the entire foreign IR module, just what you're importing.

While it's probably *possible* to integrate it into our own incr. comp. framework and have reuse "just work" it may also be good to shoehorn as close as we can to upstream incremental ThinLTO as that's how it's written to work in the sense of fixing bugs and getting future developments.

I sort of assumed that the way we'd implement incremental ThinLTO was to basically implement custom incrementality in ThinLTO passes (like upstream LLVM does) and basically just get the main incremental directory from the compilation session.

I do think we have to solve the problem here in the sense if I change one CGU that shouldn't force all CGUs which *may* import the item to be fully recompiled. Instead only those which are *known* to import the changed item should be recompiled (and even then, not fully recompiled, just post-ThinLTO-passes).