@SimonSapin 2018-07-13
> cache size already is a concern for big projects

It’s not just the incremental cache and it’s not just Rust code, but as a single data point I regularly fill the 500 GB on my desktop machine that is dedicated to compilation. Part of this is due to Cargo not having any garbage collection outside of `cargo clean`. Upgrading to a new Rust Nightly changes the hash in some directory names and can effectively double the size of `target/`.