@michaelwoerister 2018-07-13
> The description leads me to believe that in the above scenario no matter what changes in B then A is fully recompiled unconditionally. 

That's the case. The PR does NOY yet implement change tracking at the function level. Do you know if linker-based LTO computes a hash per function? I don't have the code at hand at the moment. The description here sounds like no: http://blog.llvm.org/2016/06/thinlto-scalable-and-incremental-lto.html.