@alexcrichton 2018-07-13
Aha it appears I'm misremembering! I've gone off [this list](https://github.com/llvm-mirror/llvm/blob/6ab694dd51ed0103163d44c24cfb645b1dbe9199/lib/LTO/ThinLTOCodeGenerator.cpp#L344-L385) before, and you're right in that it only looks at entire modules, not hashes of imported contents.

In light of that I think there's still a "main thing" for this PR to implement, right? Which is that we shouldn't recodegen/reoptimize entire modules, but rather only rerun the ThinLTO passes when an input changes, right? Does that make sense?