@michaelwoerister 2018-07-14
> In light of that I think there's still a "main thing" for this PR to implement, right? Which is that we shouldn't recodegen/reoptimize entire modules, but rather only rerun the ThinLTO passes when an input changes, right? Does that make sense?

Yes, that makes sense. That should put `rustc` on par with the linker plugin.